fix(lint): clear pre-existing CI debt for v0.2.0 release
53 ruff errors + 5 mypy errors that have been red on main since at
least 2026-04-29. Required to land v0.2.0 release on green CI.

Ruff (auto-fixed via `ruff check --fix --unsafe-fixes` + `ruff format`):
- I001 import sort
- B905 zip() strict= explicit
- B007 unused loop vars (renamed to _)
- E702 multiple statements on one line
- F401 unused imports
- F541 f-string without placeholders
- F841 unused local variables
- RUF021 parenthesize chained and/or
- RUF022 __all__ sorted
- RUF046 redundant int() cast
- RUF100 unused noqa directives
- UP037 quoted type annotations

Ruff (manual):
- E741 ambiguous `l`: scripts/compare_models.py:34, scripts/make_graph_figures.py:627
  renamed to `line` and `lbl` respectively

Mypy:
- src/.../agents/medical_agent.py:241 — `args: dict[str, Any]` annotation
- src/.../models/{qwen_vl,llava,internvl2}.py — `cast(torch.Tensor, outputs.logits)`
- src/.../models/llava.py:36 — `# type: ignore[attr-defined]` on
  `processor.tokenizer.eos_token_id` (LlavaNextProcessor stub gap)

Verified: 223 tests pass, ruff/format clean, mypy 5 fixed errors gone.
Pre-existing missing-types-PyYAML errors only show locally; CI installs
types-PyYAML in lint.yml so it stays clean there.

+++ src/adversarial_reasoning/attacks/apgd.py
@@ -40,7 +40,7 @@ def _checkpoints(n_iter: int) -> list[int]:
     while p[-1] < 1.0:
         nxt = p[-1] + max(p[-1] - p[-2] - 0.03, 0.06)
         p.append(min(nxt, 1.0))
-    pts = sorted({max(1, int(math.ceil(pj * n_iter))) for pj in p[1:]})
+    pts = sorted({max(1, math.ceil(pj * n_iter)) for pj in p[1:]})
     return [c for c in pts if c <= n_iter]
